一些优化和改动 - 添加mdns，使网页管理可以在固定的网址上访问 - 启用ESP32掉电复位 - 优化代码

+++ lib/RDA5807/src/RDA5807.h
@@ -754,7 +754,7 @@ public:
     void setBand3_50_65_Mode(uint8_t band3Mode);
     uint8_t getBand3Status();
 
-    int getRssi();
+    uint8_t getRssi();
 
     /**
      * @ingroup GA03
